+++ src/app/[lng]/Components/AdminContent/AdminContent.style.ts
@@ -1,0 +1,10 @@
+export const styles = {
+    box: {
+        justifyContent: "center",
+        alignItems: "center",
+        height: "100vh",
+    },
+    button: {
+        margin: "auto",
+    },
+};
