+++ src/app/[lng]/Components/AdminContent/index.ts
@@ -1,0 +1,2 @@
+import AdminContent from '@/app/[lng]/Components/AdminContent/AdminContent';
+export default AdminContent;
