+++ src/app/[lng]/admin/page.style.ts
@@ -1,14 +1,6 @@
 export const styles = {
-  container: {
-    marginTop: "20px",
-    textAlign: "center",
-  },
-  box: {
-    justifyContent: "center",
-    alignItems: "center",
-    height: "100vh",
-  },
-  button: {
-    margin: "auto",
-  },
+    container: {
+        marginTop: "20px",
+        textAlign: "center",
+    },
 };
